Move rdma_netlink compat into CMake
Detect if the distro's rdma_netlink.h is new enough, if not replace
it with the built in copy, and eliminate the two loose copies of the
header.

The built in copy is from v4.8
